groundwork for per shot fancy death effects that can control the movement of the exploding parts.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21672 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/GuidedMissleStrategy.cxx
@@ -205,7 +205,7 @@ void GuidedMissileStrategy::update(float dt)
 
   renderTimes++;
   if (puffTime < 0 )
-    puffTime =  bzfrand()*rootPuff;
+    puffTime =  (float)bzfrand()*rootPuff;
 
   // Changed: GM smoke trail, leave it every seconds, none of this per frame crap
   if (currentTime.getSeconds() - lastPuff.getSeconds() > puffTime ) {
@@ -213,7 +213,7 @@ void GuidedMissileStrategy::update(float dt)
     addShotPuff(nextPos,azimuth,elevation);
 
     // pick a new time for the next puff so it's not so orderd.
-    puffTime = bzfrand()*rootPuff;
+    puffTime = (float)bzfrand()*rootPuff;
   }
 
   // get next position
